user profile section complete. images added to thread list partial. mark solution is next

+++ routes/web.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Auth\RegisterController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\ThreadController;
 use App\Http\Controllers\CommentController;
+use App\Http\Controllers\UserProfileController;
 
 
 /*
@@ -26,30 +27,48 @@ Route::get('/', function () {
     
 })->name('home');
 
-
+// Log in routes
 Route::get('/login', [LoginController::class, 'index'])->name('login');
 Route::post('/login', [LoginController::class, 'store']);
 
+// Log out route
 Route::post('/logout', [LogoutController::class, 'store'])->name('logout');
 
+// Register routes
 Route::get('/register', [RegisterController::class, 'index'])->name('register');
 Route::post('/register', [RegisterController::class, 'store']);
 
-
-Route::resource('thread', ThreadController::class);
-
-
-Route::post('comment/create/{thread}', [CommentController::class, 'addThreadComment'])->name('threadcomment.store');
-Route::put('comment/{comment}', [CommentController::class, 'updateThreadComment'])->name('threadcomment.update');
-Route::delete('comment/{comment}', [CommentController::class, 'deleteThreadComment'])->name('threadcomment.destroy');
-
-Route::post('reply/create/{comment}', [CommentController::class, 'addReplyComment'])->name('replycomment.store');
-Route::put('reply/{comment}', [CommentController::class, 'updateReplyComment'])->name('replycomment.update');
-Route::delete('reply/{comment}', [CommentController::class, 'deleteReplyComment'])->name('replycomment.destroy');
-
+// Thread routes
+Route::get('/thread', [ThreadController::class, 'index'])->name('thread.index');
+Route::get('/thread/create', [ThreadController::class, 'create'])->name('thread.create')->middleware('auth');
+Route::post('/thread', [ThreadController::class, 'store'])->name('thread.store')->middleware('auth');
+Route::get('/thread/{thread}/edit', [ThreadController::class, 'edit'])->name('thread.edit')->middleware('auth');
+Route::get('/thread/{thread}', [ThreadController::class, 'show'])->name('thread.show');
+Route::put('/thread/{thread}', [ThreadController::class, 'update'])->name('thread.update')->middleware('auth');
+Route::delete('/thread/{thread}', [ThreadController::class, 'destroy'])->name('thread.destroy')->middleware('auth');
+Route::post('/thread/{thread}/like', [ThreadController::class, 'like'])->name('thread.like')->middleware('auth');
+Route::delete('/thread/{thread}/unlike', [ThreadController::class, 'unlike'])->name('thread.unlike')->middleware('auth');
 
 
+// Comment routes
+Route::post('comment/create/{thread}', [CommentController::class, 'addThreadComment'])->name('threadcomment.store')->middleware('auth');
+Route::put('comment/{comment}', [CommentController::class, 'updateThreadComment'])->name('threadcomment.update')->middleware('auth');
+Route::delete('comment/{comment}', [CommentController::class, 'deleteThreadComment'])->name('threadcomment.destroy')->middleware('auth');
+Route::post('/comment/{comment}/like', [CommentController::class, 'likeThreadComment'])->name('threadcomment.like')->middleware('auth');
+Route::delete('/comment/{comment}/unlike', [CommentController::class, 'unlikeThreadComment'])->name('threadcomment.unlike')->middleware('auth');
+
+// Comment reply routes
+Route::post('reply/create/{comment}', [CommentController::class, 'addReplyComment'])->name('replycomment.store')->middleware('auth');
+Route::put('reply/{comment}', [CommentController::class, 'updateReplyComment'])->name('replycomment.update')->middleware('auth');
+Route::delete('reply/{comment}', [CommentController::class, 'deleteReplyComment'])->name('replycomment.destroy')->middleware('auth');
+// Route::post('/reply/{comment}/like', [CommentController::class, 'likeReplyComment'])->name('replycomment.like')->middleware('auth');
+
+// User profile routes
+Route::get('/user/{user}/profile', [UserProfileController::class, 'index'])->name('userprofile')->middleware('auth');
+
+// Dashboard route
 Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
+
 
 Route::get('/latest', function () {
     return view('latest.index');
